ip2gid: Move server code to a separate file
Move server code to a separate file server.c.
